16-11-22

+++ main.py
@@ -1,6 +1,6 @@
 import tkinter.filedialog as tk
 import json
-import random
+from random import choice
 
 import pandas as pd
 import telebot
@@ -11,7 +11,7 @@ token, admins, file, secrets_words = config['token'], config['admins'], config['
 map(int, admins)
 
 db = pd.read_excel(file)
-bot = telebot.TeleBot(token)
+bot = telebot.TeleBot(token, 'MARKDOWN')
 
 registered_users = set(db['ФИО'].to_list())
 active_secret_word = {}
@@ -34,11 +34,17 @@ def admin (message):
             keyboard = types.InlineKeyboardMarkup()
             keyboard.add(types.InlineKeyboardButton(text = 'Сгенерировать новое', callback_data = 'admin|gen'))
             keyboard.add(types.InlineKeyboardButton(text = 'Деактивировать слово', callback_data = 'admin|delete'))
-            bot.send_message(message.from_user.id, f'Есть активное сектное слово: {list(active_secret_word.keys())[0]}, для {list(active_secret_word.values())[0]} класса', reply_markup=keyboard)
+            bot.send_message(message.from_user.id, f'Есть активное сектное слово: *{list(active_secret_word.keys())[0]}*, для *{list(active_secret_word.values())[0]}* класса', reply_markup=keyboard)
         else:
             keyboard = types.InlineKeyboardMarkup()
             keyboard.add(types.InlineKeyboardButton(text = 'Сгенерировать слово', callback_data = 'admin|gen'))
             bot.send_message(message.from_user.id, f'Активных секретных слов нет', reply_markup=keyboard)
+
+
+@bot.message_handler(content_types=['text'])
+def get_text_messages(message):
+    if message.text.lower() == list(active_secret_word.keys())[0]:
+        bot.send_message(message.from_user.id, list(active_secret_word.values())[0])
 
 
 @bot.callback_query_handler(func=lambda call: True)
@@ -46,11 +52,23 @@ def callback_worker(call):
     # print(call)
     # print(call.data)
     call_data = call.data.split('|')
-    if call_data[0] == 'admin':
+    if call_data[0] == 'admin' and call.from_user.id in admins:
         bot.delete_message(call.from_user.id, call.message.id)
         if call_data[1] == 'gen':
+            keyboard = types.InlineKeyboardMarkup()
+            tmp = list(set(db['Класс'].to_list()))
+            # tmp.sort()
+            for i in tmp:
+                keyboard.add(types.InlineKeyboardButton(text = i, callback_data = f'admin|create|{i}'))
+            bot.send_message(call.from_user.id, f'Выберите класс для которого сгенерируется слово', reply_markup=keyboard)
+        elif call_data[1] == 'create':
             active_secret_word.clear()
-            active_secret_word.update([[]])
+            active_secret_word.update([[choice(secrets_words), call_data[2]]])
+            admin (call)
+        elif call_data[1] == 'delete':
+            active_secret_word.clear()
+            admin (call)
+
